Api made restful.
